creates models module for gather metadata Settings
